requirements.txt
streamlit
fastapi
uvicorn
numpy
pandas
openpyxl
python-multipart
scipy
Pillow
zstandard
